added help component
