+++ layouts/event.vue
@@ -277,196 +277,20 @@
       </v-container>
     </v-main>
 
-    <div class="help-section">
-      <v-btn
-        class="mx-2 help-button"
-        fab
-        small
-        outlined
-        color="primary"
-        @click="helpDialog = !helpDialog"
-      >
-        <v-icon>
-          mdi-help
-        </v-icon>
-      </v-btn>
-      <v-dialog
-        v-model="helpDialog"
-        content-class="help-box"
-        transition="dialog-bottom-transition"
-      >
-        <v-card>
-          <div class="text-center pa-2 help-top-section positionRelative">
-            <div class="robo-img mt-2">
-              <v-img :src="$config.cdnUri + 'robo-white.png'"></v-img>
-            </div>
-            <div class="headline ma-2 white--text">
-              <i18n path="Common.HowHelp" />
-            </div>
-            <span class="positionAbsolute help-close">
-              <v-btn icon color="white" @click="close()">
-                <v-icon>mdi-close</v-icon>
-              </v-btn>
-            </span>
-            <span v-if="!helpForm" class="positionAbsolute help-back">
-              <v-btn icon color="white" @click="helpForm = true">
-                <v-icon>fa-arrow-left</v-icon>
-              </v-btn>
-            </span>
-          </div>
-          <div v-if="helpForm">
-            <a
-              href="https://event.bitpod.io/admin/docs/HelpCenter/"
-              target="_blank"
-            >
-              <div class="d-flex pt-4 px-2">
-                <div class="chanel-icon d-flex ma-4 mr-6">
-                  <i
-                    class="fa fa-flag fs-18 d-flex align-center justify-center"
-                    style="background: #0cb14b;"
-                  ></i>
-                </div>
-                <div
-                  class="chanel-right d-flex flex-column justify-center align-start"
-                >
-                  <div class="chanel-right-head">
-                    <i18n path="Common.UserGuide" />
-                  </div>
-                  <div class="chanel-right-info body-2">
-                    <i18n path="Common.UserGuideInfo" />
-                  </div>
-                </div>
-                <div class="chanel-more d-flex align-center justify-center">
-                  <i class="fa fa-chevron-right fs-22"></i>
-                </div>
-              </div>
-            </a>
-            <a
-              href="https://event.bitpod.io/admin/docs/HelpCenter/Introduction/FAQs/views/FAQs"
-              target="_blank"
-            >
-              <div class="d-flex pt-4 px-2">
-                <div class="chanel-icon d-flex ma-4 mr-6">
-                  <i
-                    class="fa fa-support fs-18 d-flex align-center justify-center"
-                    style="background: #f4b400;"
-                  ></i>
-                </div>
-                <div
-                  class="chanel-right d-flex flex-column justify-center align-start"
-                >
-                  <div class="chanel-right-head">
-                    <i18n path="Common.Faqs" />
-                  </div>
-                  <div class="chanel-right-info body-2">
-                    <i18n path="Common.FaqsInfo" />
-                  </div>
-                </div>
-                <div class="chanel-more d-flex align-center justify-center">
-                  <i class="fa fa-chevron-right fs-22"></i>
-                </div>
-              </div>
-            </a>
-            <a
-              href="https://www.youtube.com/channel/UCPE_a4LrIlP2RL04P92Bl-Q"
-              target="_blank"
-            >
-              <div class="d-flex pt-4 px-2">
-                <div class="chanel-icon d-flex ma-4 mr-6">
-                  <i
-                    class="fa fa-youtube fs-18 d-flex align-center justify-center"
-                    style="background: #f25955;"
-                  ></i>
-                </div>
-                <div
-                  class="chanel-right d-flex flex-column justify-center align-start"
-                >
-                  <div class="chanel-right-head">
-                    <i18n path="Common.HelpVideos" />
-                  </div>
-                  <div class="chanel-right-info body-2">
-                    <i18n path="Common.HelpVideosInfo" />
-                  </div>
-                </div>
-                <div class="chanel-more d-flex align-center justify-center">
-                  <i class="fa fa-chevron-right fs-22"></i>
-                </div>
-              </div>
-            </a>
-            <v-card-actions>
-              <v-btn
-                block
-                depressed
-                color="primary"
-                class="my-3"
-                @click="helpForm = false"
-              >
-                <i18n path="Common.ContactSupport" />
-              </v-btn>
-            </v-card-actions>
-          </div>
-          <div v-else>
-            <v-col class="col-12 col-md-12 py-0 mt-4">
-              <v-text-field
-                :value="this.$auth.$state.user.data.name"
-                label="Name"
-                outlined
-                disabled
-                dense
-              ></v-text-field>
-            </v-col>
-            <v-col class="col-12 col-md-12 py-0">
-              <v-text-field
-                :value="this.$auth.$state.user.data.email"
-                label="Email"
-                outlined
-                disabled
-                dense
-              ></v-text-field>
-            </v-col>
-            <v-col class="col-12 col-md-12 py-0">
-              <v-textarea
-                v-model="helpMessage"
-                label="Message"
-                outlined
-                no-resize
-                rows="2"
-                row-height="25"
-                height="100"
-              ></v-textarea>
-            </v-col>
-            <v-card-actions>
-              <v-btn
-                block
-                depressed
-                color="primary"
-                class="mb-3 mt-0"
-                @click="onSave"
-              >
-                <i18n path="Common.LeaveMessage" />
-              </v-btn>
-            </v-card-actions>
-          </div>
-        </v-card>
-      </v-dialog>
-      <v-snackbar v-model="snackbar" :timeout="timeout" :top="true">
-        <div class="fs-16 text-center">
-          {{ snackbarText }}
-        </div>
-      </v-snackbar>
-    </div>
+    <Help />
   </v-app>
 </template>
 
 <script>
 import OrgnaizationList from '~/components/common/organization-list'
 import AppDrawer from '~/components/common/app-drawer'
-import nuxtconfig from '~/nuxt.config'
+import Help from '~/components/common/help'
 export default {
   middleware: ['auth', 'authorization'],
   components: {
     OrgnaizationList,
     AppDrawer,
+    Help,
   },
   props: {
     source: { type: String, default: '' },
@@ -487,17 +311,6 @@ export default {
     triggerReset: false,
     triggerRecEventReset: false,
     activeClass: ' v-list-item--active',
-    helpDialog: false,
-    helpForm: true,
-    formData: {
-      Name: '',
-      Email: '',
-      Message: '',
-    },
-    helpMessage: '',
-    snackbar: false,
-    timeout: '4000',
-    snackbarText: '',
     items: [
       {
         icon: 'fa fa-grid',
@@ -579,34 +392,6 @@ export default {
       }, false)
         ? this.activeClass
         : ''
-    },
-    async onSave() {
-      this.formData.Name = this.$auth.$state.user.data.name
-      this.formData.Email = this.$auth.$state.user.data.email
-      this.formData.Message = this.helpMessage
-      try {
-        const url = `https://${this.$config.axios.crmUrl}${nuxtconfig.axios.apiEndpoint}`
-        const res = await this.$axios.$patch(`${url}Leads`, {
-          ...this.formData,
-        })
-        if (res) {
-          this.helpDialog = false
-          this.helpForm = true
-          this.onReset()
-          this.snackbarText = this.$t('Messages.Success.ThanksHelp')
-          this.snackbar = true
-        }
-      } catch (e) {
-        console.log('Error', e)
-      }
-    },
-    onReset() {
-      this.helpMessage = ''
-    },
-    close() {
-      this.helpDialog = false
-      this.helpForm = true
-      this.helpMessage = ''
     },
   },
 }
